deploy/requirements.txt
# PyTorch CPU only (GPU 불필요, Lambda용)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# YOLOv8
ultralytics

# Segmentation
segmentation-models-pytorch

# RAG
faiss-cpu

# Image processing
Pillow
numpy
opencv-python-headless

# AWS
boto3

# Utilities
scipy
